Refactor command handling by moving server start logic from index.ts to a new file under /server, startCommand.ts
